requirements.txt
Django>=3.2,<4.0
djangorestframework
celery
django-celery-beat
django-celery-results
django-channels
django-two-factor-auth>=1.15.0
django-otp>=1.1.3
phonenumbers>=8.13.0
qrcode>=7.3
channels
paramiko
psycopg2-binary
gunicorn
pytest
pytest-django
pytest-asyncio
selenium
psutil
paramiko
redis